# web/lib/api/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "StoreOps.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/admin/products": {
      "get": {
        "tags": [
          "AdminProducts"
        ],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "schema": {
              "maxLength": 100,
              "minLength": 0,
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "default": "active",
              "$ref": "#/components/schemas/ProductStatus"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "maximum": 10000,
              "minimum": 1,
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "maximum": 100,
              "minimum": 1,
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PagedOfAdminProductResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedOfAdminProductResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedOfAdminProductResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "AdminProducts"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ProductRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProductResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProductResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProductResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/products/{id}": {
      "get": {
        "tags": [
          "AdminProducts"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProductResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProductResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProductResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "AdminProducts"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProductResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProductResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProductResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/products/{id}/deal": {
      "put": {
        "tags": [
          "AdminProducts"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DealRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DealRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DealRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProductResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProductResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProductResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "AdminProducts"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProductResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProductResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProductResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/products/{id}/archive": {
      "post": {
        "tags": [
          "AdminProducts"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProductResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProductResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProductResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/products/{id}/restore": {
      "post": {
        "tags": [
          "AdminProducts"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProductResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProductResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProductResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/products/bulk": {
      "post": {
        "tags": [
          "AdminProducts"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductBulkRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ProductBulkRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ProductBulkRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/BulkResultOfint"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkResultOfint"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/BulkResultOfint"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/products": {
      "get": {
        "tags": [
          "Catalog"
        ],
        "parameters": [
          {
            "name": "categoryId",
            "in": "query",
            "schema": {
              "type": "array",
              "items": {
                "type": "integer",
                "format": "int32"
              }
            }
          },
          {
            "name": "search",
            "in": "query",
            "schema": {
              "maxLength": 100,
              "minLength": 0,
              "type": "string"
            }
          },
          {
            "name": "onDeal",
            "in": "query",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "default": "newest",
              "$ref": "#/components/schemas/ProductSort"
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "maximum": 10000,
              "minimum": 1,
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "maximum": 50,
              "minimum": 1,
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PagedOfProductResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedOfProductResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedOfProductResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/products/{id}": {
      "get": {
        "tags": [
          "Catalog"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/products/{id}/related": {
      "get": {
        "tags": [
          "Catalog"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "maximum": 12,
              "minimum": 1,
              "type": "integer",
              "format": "int32",
              "default": 4
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProductResponse"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdminProductResponse": {
        "required": [
          "id",
          "name",
          "description",
          "categoryId",
          "categoryName",
          "priceCents",
          "compareAtPriceCents",
          "dealDescription",
          "stock",
          "imageKey",
          "imageUrl",
          "createdAtUtc",
          "updatedAtUtc",
          "archivedAtUtc",
          "rowVersion"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "categoryId": {
            "type": "integer",
            "format": "int32"
          },
          "categoryName": {
            "type": "string"
          },
          "priceCents": {
            "type": "integer",
            "format": "int32"
          },
          "compareAtPriceCents": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "dealDescription": {
            "type": [
              "null",
              "string"
            ]
          },
          "stock": {
            "type": "integer",
            "format": "int32"
          },
          "imageKey": {
            "type": "string"
          },
          "imageUrl": {
            "type": "string"
          },
          "createdAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "archivedAtUtc": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "rowVersion": {
            "type": "string",
            "format": "byte"
          }
        }
      },
      "BulkFailureOfint": {
        "required": [
          "id",
          "code",
          "message"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "BulkResultOfint": {
        "required": [
          "succeeded",
          "failed"
        ],
        "type": "object",
        "properties": {
          "succeeded": {
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            }
          },
          "failed": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BulkFailureOfint"
            }
          }
        }
      },
...
      "DealRequest": {
        "type": "object",
        "properties": {
          "dealPriceCents": {
            "maximum": 10000000,
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "description": {
            "maxLength": 200,
            "minLength": 0,
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "PagedOfAdminProductResponse": {
        "required": [
          "items",
          "page",
          "pageSize",
          "totalCount",
          "totalPages"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminProductResponse"
            }
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "PagedOfProductResponse": {
        "required": [
          "items",
          "page",
          "pageSize",
          "totalCount",
          "totalPages"
        ],
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductResponse"
            }
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "ProductBulkAction": {
        "enum": [
          "archive",
          "end_deal",
          "move"
        ]
      },
      "ProductBulkRequest": {
        "required": [
          "ids"
        ],
        "type": "object",
        "properties": {
          "ids": {
            "maxItems": 100,
            "minItems": 1,
            "type": "array",
            "items": {
              "type": "integer",
              "format": "int32"
            }
          },
          "action": {
            "$ref": "#/components/schemas/ProductBulkAction"
          },
          "categoryId": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          }
        }
      },
      "ProductRequest": {
        "required": [
          "name",
          "description",
          "imageKey"
        ],
        "type": "object",
        "properties": {
          "name": {
            "maxLength": 120,
            "minLength": 0,
            "type": "string"
          },
          "description": {
            "maxLength": 2000,
            "minLength": 0,
            "type": "string"
          },
          "categoryId": {
            "maximum": 2147483647,
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "priceCents": {
            "maximum": 10000000,
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "stock": {
            "maximum": 100000,
            "minimum": 0,
            "type": "integer",
            "format": "int32"
          },
          "imageKey": {
            "maxLength": 300,
            "minLength": 0,
            "type": "string"
          }
        }
      },
      "ProductResponse": {
        "required": [
          "id",
          "name",
          "description",
          "priceCents",
          "compareAtPriceCents",
          "dealDescription",
          "stock",
          "categoryId",
          "categoryName",
          "imageUrl"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "priceCents": {
            "type": "integer",
            "format": "int32"
          },
          "compareAtPriceCents": {
            "type": [
              "null",
              "integer"
            ],
            "format": "int32"
          },
          "dealDescription": {
            "type": [
              "null",
              "string"
            ]
          },
          "stock": {
            "type": "integer",
            "format": "int32"
          },
          "categoryId": {
            "type": "integer",
            "format": "int32"
          },
          "categoryName": {
            "type": "string"
          },
          "imageUrl": {
            "type": "string"
          }
        }
      },
      "ProductSort": {
        "enum": [
          "newest",
          "oldest",
          "cheapest",
          "priciest"
        ]
      },
      "ProductStatus": {
        "enum": [
          "active",
          "archived",
          "all"
        ]
      },
...
      "UpdateProductRequest": {
        "required": [
          "rowVersion",
          "name",
          "description",
          "imageKey"
        ],
        "type": "object",
        "properties": {
          "rowVersion": {
            "maxLength": 8,
            "minLength": 8,
            "type": "string",
            "format": "byte"
          },
          "name": {
            "maxLength": 120,
            "minLength": 0,
            "type": "string"
          },
          "description": {
            "maxLength": 2000,
            "minLength": 0,
            "type": "string"
          },
          "categoryId": {
            "maximum": 2147483647,
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "priceCents": {
            "maximum": 10000000,
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "stock": {
            "maximum": 100000,
            "minimum": 0,
            "type": "integer",
            "format": "int32"
          },
          "imageKey": {
            "maxLength": 300,
            "minLength": 0,
            "type": "string"
          }
        }
      },
...
    }
  },
...
}